requirements.txt
streamlit
openai
altair
pysqlite3
bcrypt
python-dotenv
